refactor: 템플릿 메소드 패턴을 사용하여 중복 코드 대응
